requirements.txt
annotated-types==0.7.0
anyio==4.6.2.post1
bcrypt==4.2.0
Brotli==1.0.9
certifi==2021.10.8
charset-normalizer==2.0.10
click==8.1.7
colorama==0.4.4
comtypes==1.2.1
contourpy==1.3.0
cycler==0.12.1
discum==1.4.0
ecdsa==0.19.0
exceptiongroup==1.2.2
fastapi==0.115.4
filetype==1.0.9
fonttools==4.54.1
h11==0.14.0
httpcore==1.0.6
httpx==0.27.2
idna==3.3
importlib_resources==6.4.5
iniconfig==2.0.0
kiwisolver==1.4.7
matplotlib==3.9.2
MouseInfo==0.1.3
numpy==1.21.4
opencv-python==4.5.4.60
packaging==24.1
passlib==1.7.4
Pillow==9.0.0
pluggy==1.5.0
pyasn1==0.6.1
PyAutoGUI==0.9.54
pydantic==2.9.2
pydantic_core==2.23.4
PyGetWindow==0.0.9
pygrabber==0.1
PyMsgBox==1.0.9
pyparsing==3.1.4
pyperclip==1.8.2
PyRect==0.2.0
PyScreeze==0.1.28
pytest==8.3.3
pytest-asyncio==0.24.0
python-dateutil==2.9.0.post0
python-jose==3.3.0
python-multipart==0.0.17
pytweening==1.0.7
pywin32==303
requests==2.27.1
requests-toolbelt==0.9.1
rsa==4.9
six==1.16.0
sniffio==1.3.1
starlette==0.41.2
tomli==2.0.2
typing_extensions==4.12.2
ua-parser==0.10.0
urllib3==1.26.8
uvicorn==0.32.0
websocket-client==0.59.0
zipp==3.20.2
